requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sentence-transformers==2.6.1
pydantic==2.7.1
python-dotenv==1.0.1
numpy==1.26.4
torch>=2.5.0
torchvision>=0.21.0
torchaudio>=2.5.0
transformers==4.38.2
networkx==3.2.1
python-multipart>=0.0.7
requests==2.31.0
httpx==0.27.0
# Removed mistralai package since we're using OpenRouter directly
pydantic-settings==2.2.1
tqdm==4.66.1
markdown==3.5.2
SQLAlchemy>=2.0.0
# PDF processing dependencies
pdfplumber==0.10.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
scikit-learn==1.4.0
asyncpg>=0.27.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
